Move remaining host-agnostic editor pieces into ag-website-shared
Moves the theme-builder editor components/chrome that had no
grid-specific content into ag-website-shared/src/components/theme-builder:
ErrorBoundary, Checkbox (now using classnames instead of grid's local
combineClassNames), Tooltip, FormField, dropdown-shared, icons,
ValueEditorProps, BorderStyleValueEditor, ColorSchemeValueEditor,
FontWeightValueEditor, BorderValueEditor, PartEditor, and
CssValueEditor.

Also deletes editors/Input.tsx, FormattedInput.tsx and
formatted-input-global-styles.css - dead since the sidebar migration
moved their only remaining consumers onto the shared versions.

Rewires all consumers (ParamEditor, BordersEditor, ImageValueEditor,
FontFamilyValueEditor, ColorValueEditor, LengthValueEditor,
ScaleValueEditor, EditorPanel, GridPreview, GridConfigDropdown) onto
the shared locations.

Left in place (grid-specific content, not generic components):
BordersEditor's border list, FontFamilyValueEditor's curated font
list, ImageValueEditor's docs link, and ParamEditor's dispatcher
(its ValueEditorProps/ParamModel-coupled contract needs reconciling
with the shared leaf editors' plain value/onChange shape first).

Verified: tsc identical to baseline, eslint/prettier clean, 40/40
tests pass, and manually in the running dev server (borders dropdown,
icon-set selector, and grid-features checkboxes all render and behave
correctly).
