+++ documentation/ag-grid-docs/src/components/theme-builder/components/editors/BordersEditor.tsx
@@ -1,3 +1,11 @@
+import { withErrorBoundary } from '@ag-website-shared/components/theme-builder/ErrorBoundary';
+import { FormField } from '@ag-website-shared/components/theme-builder/FormField';
+import {
+    SharedContent,
+    SharedIndicator,
+    SharedItem,
+    SharedTrigger,
+} from '@ag-website-shared/components/theme-builder/dropdown-shared';
 import { ParamModel, useParamAtom } from '@ag-website-shared/theming/ParamModel';
 import { paramValueToCss } from '@ag-website-shared/theming/api';
 import { useRenderedTheme } from '@ag-website-shared/theming/rendered-theme';
@@ -6,10 +14,6 @@ import type { ThemeParam } from '@ag-website-shared/theming/utils';
 import { Checkmark, ChevronDown } from '@carbon/icons-react';
 import styled from '@emotion/styled';
 import * as RadixDropdown from '@radix-ui/react-dropdown-menu';
-
-import { withErrorBoundary } from '../general/ErrorBoundary';
-import { FormField } from './FormField';
-import { SharedContent, SharedIndicator, SharedItem, SharedTrigger } from './dropdown-shared';
 
 const borders: [ThemeParam, string][] = [
     ['wrapperBorder', 'Around grid'],
