+++ documentation/ag-grid-docs/src/components/theme-builder/components/editors/ColorValueEditor.tsx
@@ -1,7 +1,6 @@
 import { ColorPicker } from '@ag-website-shared/components/theme-builder/ColorPicker';
+import type { ValueEditorProps } from '@ag-website-shared/components/theme-builder/ValueEditorProps';
 import { type ColorValue, paramValueToCss } from '@ag-website-shared/theming/api';
-
-import { type ValueEditorProps } from './ValueEditorProps';
 
 export const ColorValueEditor = ({ param, value, onChange }: ValueEditorProps<ColorValue>) => (
     <ColorPicker
