+++ documentation/ag-grid-docs/src/components/theme-builder/components/editors/LengthValueEditor.tsx
@@ -1,7 +1,6 @@
 import { LengthInput } from '@ag-website-shared/components/theme-builder/LengthInput';
+import type { ValueEditorProps } from '@ag-website-shared/components/theme-builder/ValueEditorProps';
 import { type LengthValue, paramValueToCss } from '@ag-website-shared/theming/api';
-
-import { type ValueEditorProps } from './ValueEditorProps';
 
 export const LengthValueEditor = ({
     param,
