--- documentation/ag-grid-docs/src/components/theme-builder/components/editors/formatted-input-global-styles.css
@@ -1,3 +1,0 @@
-.force-resize-cursor * {
-    cursor: ew-resize !important;
-}
