+++ external/ag-website-shared/src/components/theme-builder/BorderValueEditor.tsx
@@ -1,6 +1,5 @@
-import { type BorderValue } from '@ag-website-shared/theming/api';
-
-import { Checkbox } from '../general/Checkbox';
+import { type BorderValue } from '../../theming/api';
+import { Checkbox } from './Checkbox';
 import type { ValueEditorProps } from './ValueEditorProps';
 
 export const BorderValueEditor = ({ value, onChange }: ValueEditorProps<BorderValue>) => {
