+++ external/ag-website-shared/src/components/theme-builder/ColorSchemeValueEditor.tsx
@@ -1,6 +1,6 @@
 import { Select } from '@ag-website-shared/components/select/Select';
-import { type ColorSchemeValue } from '@ag-website-shared/theming/api';
 
+import { type ColorSchemeValue } from '../../theming/api';
 import type { ValueEditorProps } from './ValueEditorProps';
 
 export const ColorSchemeValueEditor = ({ value, onChange }: ValueEditorProps<ColorSchemeValue>) => {
