+++ external/ag-website-shared/src/components/theme-builder/ValueEditorProps.ts
@@ -1,5 +1,6 @@
-import type { ParamModel } from '@ag-website-shared/theming/ParamModel';
 import type { ReactNode } from 'react';
+
+import type { ParamModel } from '../../theming/ParamModel';
 
 export type ValueEditorProps<T> = {
     param: ParamModel<T>;
